fix: simplify UTF-8 chunk reading and pass CSV lines as array to workers
- Simplify read_file_chunk UTF-8 boundary handling
- Skip entire chunk on invalid UTF-8 to prevent infinite loop
- Move splitCsvLines to csvLineParser.ts for shared use
- Pass batch lines as string[] to workers instead of joined string
- Fix quoted fields with embedded newlines being split incorrectly
- Use splitCsvLines in ImportFile header validation
